# backend/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API для управления складом товаров с ролевой моделью доступа",
    "title": "MarketEase API",
    "contact": {},
    "version": "2.0"
  },
  "host": "localhost:8080",
  "basePath": "/api",
  "paths": {
    "/admin/assign-role": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Только для director",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["admin"],
        "summary": "Изменить роль пользователя",
        "parameters": [
          {
            "description": "Данные для изменения роли",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/controllers.AssignRoleInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
    "/admin/ban-user": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Только для director",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["admin"],
        "summary": "Заблокировать пользователя",
        "parameters": [
          {
            "description": "ID пользователя",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/controllers.UserIDInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
    "/admin/delete-user": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Только для director",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["admin"],
        "summary": "Удалить пользователя",
        "parameters": [
          {
            "description": "ID пользователя",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/controllers.UserIDInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
    "/admin/deleted-users": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Только для director",
        "produces": ["application/json"],
        "tags": ["admin"],
        "summary": "Получить список удаленных пользователей",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/models.User"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
    "/admin/restore-user": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Только для director",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["admin"],
        "summary": "Восстановить удаленного пользователя",
        "parameters": [
          {
            "description": "ID пользователя",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/controllers.UserIDInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
    "/admin/unban-user": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Только для director",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["admin"],
        "summary": "Разблокировать пользователя",
        "parameters": [
          {
            "description": "ID пользователя",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/controllers.UserIDInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
    "/admin/users": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Только для director",
        "produces": ["application/json"],
        "tags": ["admin"],
        "summary": "Получить список всех пользователей",
        "parameters": [
          {
            "type": "boolean",
            "description": "Включать удаленных пользователей",
            "name": "includeDeleted",
            "in": "query"
          },
          {
            "type": "boolean",
            "description": "Включать заблокированных пользователей",
            "name": "includeBanned",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/models.User"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "controllers.AssignRoleInput": {
      "type": "object",
      "required": ["role", "userId"],
      "properties": {
        "role": {
          "type": "string",
          "example": "manager"
        },
        "userId": {
          "type": "string",
          "example": "e7b1c2f1-567b-4b14-a8d7-cd5b08bfc9d0"
        }
      }
    },
...
    "controllers.UserIDInput": {
      "type": "object",
      "required": ["userId"],
      "properties": {
        "userId": {
          "type": "string",
          "example": "e7b1c2f1-567b-4b14-a8d7-cd5b08bfc9d0"
        }
      }
    },
    "models.ErrorResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Произошла ошибка"
        }
      }
    },
...
    "models.User": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string"
        },
        "deletedAt": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "example": "john@example.com"
        },
        "id": {
          "type": "string",
          "example": "e7b1c2f1-567b-4b14-a8d7-cd5b08bfc9d0"
        },
        "role": {
          "type": "string",
          "example": "viewer"
        },
        "status": {
          "type": "string",
          "example": "active"
        },
        "updatedAt": {
          "type": "string"
        },
        "username": {
          "type": "string",
          "example": "john_doe"
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}